+++ packages/api-client/src/APIClient.ts
@@ -44,8 +44,9 @@ import {GiphyAPI} from './giphy/';
 import {HttpClient} from './http/';
 import {NotificationAPI} from './notification/';
 import {ObfuscationUtil} from './obfuscation/';
+import {ServiceProviderAPI} from './serviceProvider';
 import {SelfAPI} from './self/';
-import {WebSocketClient} from './tcp/';
+import {OnConnect, WebSocketClient} from './tcp/';
 import {
   TeamConversationAPI,
   FeatureAPI,
@@ -95,6 +96,7 @@ export class APIClient extends EventEmitter {
   public giphy: {api: GiphyAPI};
   public notification: {api: NotificationAPI};
   public self: {api: SelfAPI};
+  public serviceProvider: {api: ServiceProviderAPI};
   public teams: {
     conversation: {api: TeamConversationAPI};
     feature: {api: FeatureAPI};
@@ -155,7 +157,6 @@ export class APIClient extends EventEmitter {
       http: httpClient,
       ws: webSocket,
     };
-
     this.account = {
       api: new AccountAPI(this.transport.http),
     };
@@ -186,7 +187,9 @@ export class APIClient extends EventEmitter {
     this.self = {
       api: new SelfAPI(this.transport.http),
     };
-
+    this.serviceProvider = {
+      api: new ServiceProviderAPI(this.transport.http),
+    };
     this.teams = {
       conversation: {
         api: new TeamConversationAPI(this.transport.http),
@@ -219,7 +222,6 @@ export class APIClient extends EventEmitter {
         api: new TeamAPI(this.transport.http),
       },
     };
-
     this.user = {
       api: new UserAPI(this.transport.http),
     };
@@ -280,8 +282,8 @@ export class APIClient extends EventEmitter {
     delete this.context;
   }
 
-  public connect(onBeforeConnect?: () => Promise<void>): Promise<WebSocketClient> {
-    return this.transport.ws.connect(this.context?.clientId, onBeforeConnect);
+  public connect(onConnect?: OnConnect): Promise<WebSocketClient> {
+    return this.transport.ws.connect(this.context?.clientId, onConnect);
   }
 
   private createContext(userId: string, clientType: ClientType, clientId?: string): Context {