+++ packages/core/src/main/Account.ts
@@ -17,20 +17,13 @@
  *
  */
 
-import {APIClient} from '@wireapp/api-client';
-import {RegisterData} from '@wireapp/api-client/dist/auth';
-import {
-  AUTH_COOKIE_KEY,
-  AUTH_TABLE_NAME,
-  Context,
-  Cookie,
-  CookieStore,
-  LoginData,
-} from '@wireapp/api-client/dist/auth/';
-import {ClientType, RegisteredClient} from '@wireapp/api-client/dist/client/';
-import * as Events from '@wireapp/api-client/dist/event';
 import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
-import {WebSocketClient} from '@wireapp/api-client/dist/tcp/';
+import {APIClient} from '@wireapp/api-client';
+import {RegisterData} from '@wireapp/api-client/src/auth';
+import {AUTH_COOKIE_KEY, AUTH_TABLE_NAME, Context, Cookie, CookieStore, LoginData} from '@wireapp/api-client/src/auth/';
+import {ClientType, RegisteredClient} from '@wireapp/api-client/src/client/';
+import * as Events from '@wireapp/api-client/src/event';
+import {WebSocketClient} from '@wireapp/api-client/src/tcp/';
 import * as cryptobox from '@wireapp/cryptobox';
 import {CRUDEngine, MemoryEngine, error as StoreEngineError} from '@wireapp/store-engine';
 import {EventEmitter} from 'events';