+++ packages/core/src/main/conversation/ConversationService.ts
@@ -17,6 +17,7 @@
  *
  */
 
+import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
 import {APIClient} from '@wireapp/api-client';
 import {
   Conversation,
@@ -27,11 +28,10 @@ import {
   NewOTRMessage,
   OTRRecipients,
   UserClients,
-} from '@wireapp/api-client/dist/conversation/';
-import {CONVERSATION_TYPING, ConversationMemberUpdateData} from '@wireapp/api-client/dist/conversation/data/';
-import {ConversationMemberLeaveEvent} from '@wireapp/api-client/dist/event/';
-import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
-import {UserPreKeyBundleMap} from '@wireapp/api-client/dist/user/';
+} from '@wireapp/api-client/src/conversation/';
+import {CONVERSATION_TYPING, ConversationMemberUpdateData} from '@wireapp/api-client/src/conversation/data/';
+import {ConversationMemberLeaveEvent} from '@wireapp/api-client/src/event/';
+import {UserPreKeyBundleMap} from '@wireapp/api-client/src/user/';
 import {
   Asset,
   ButtonAction,