+++ packages/commons/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.9](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.8...@wireapp/commons@3.8.9) (2020-10-22)
+
+
+### Bug Fixes
+
+* **commons,copy-config:** Publish .d.ts files to npm ([#3291](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/issues/3291)) ([ad3ed3c](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/commit/ad3ed3c5400e9be02f2671aaa43deb0e8d884ca9))
+
+
+
+
+
+## [3.8.8](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.7...@wireapp/commons@3.8.8) (2020-10-19)
+
+**Note:** Version bump only for package @wireapp/commons
+
+
+
+
+
+## [3.8.7](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.6...@wireapp/commons@3.8.7) (2020-10-13)
+
+
+### Bug Fixes
+
+* **commons:** Legacy support for encoded streams in Chrome 83 & 84 ([#3262](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/issues/3262)) ([11257c1](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/commit/11257c10218ffdea9d1d684e5a23b3294ec3de1a))
+
+
+
+
+
+## [3.8.6](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.5...@wireapp/commons@3.8.6) (2020-10-08)
+
+**Note:** Version bump only for package @wireapp/commons
+
+
+
+
+
+## [3.8.5](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.4...@wireapp/commons@3.8.5) (2020-10-08)
+
+
+### Bug Fixes
+
+* **commons:** Remove check for deprecated 'createEncodedVideoStreams' ([#3225](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/issues/3225)) ([ac52616](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/commit/ac526166b3862d56fc01213a65bddfa9285e0083))
+
+
+
+
+
+## [3.8.4](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.3...@wireapp/commons@3.8.4) (2020-10-05)
+
+
+### Bug Fixes
+
+* **commons:** Ensure value in getOSFamily ([#3248](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/issues/3248)) ([78de3ae](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/commit/78de3aed22bbf4a1761e6b59882bce367cdac28b))
+
+
+
+
+
+## [3.8.3](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.2...@wireapp/commons@3.8.3) (2020-09-28)
+
+**Note:** Version bump only for package @wireapp/commons
+
+
+
+
+
+## [3.8.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.1...@wireapp/commons@3.8.2) (2020-09-21)
+
+**Note:** Version bump only for package @wireapp/commons
+
+
+
+
+
+## [3.8.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.8.0...@wireapp/commons@3.8.1) (2020-09-14)
+
+**Note:** Version bump only for package @wireapp/commons
+
+
+
+
+
 # [3.8.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/commons/compare/@wireapp/commons@3.7.2...@wireapp/commons@3.8.0) (2020-09-01)