+++ packages/priority-queue/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.10](https://github.com/wireapp/wire-web-packages/tree/master/packages/priority-queue/compare/@wireapp/priority-queue@1.6.9...@wireapp/priority-queue@1.6.10) (2020-10-15)
+
+**Note:** Version bump only for package @wireapp/priority-queue
+
+
+
+
+
+## [1.6.9](https://github.com/wireapp/wire-web-packages/tree/master/packages/priority-queue/compare/@wireapp/priority-queue@1.6.8...@wireapp/priority-queue@1.6.9) (2020-09-21)
+
+**Note:** Version bump only for package @wireapp/priority-queue
+
+
+
+
+
+## 1.6.8 (2020-09-14)
+
+**Note:** Version bump only for package @wireapp/priority-queue
+
+
+
+
+
 ## [1.6.7](https://github.com/wireapp/wire-web-packages/tree/master/packages/priority-queue/compare/@wireapp/priority-queue@1.6.6...@wireapp/priority-queue@1.6.7) (2020-08-11)
 
 **Note:** Version bump only for package @wireapp/priority-queue